moved alerts code to backend
