--- alerts/notifications.js
@@ -1,18 +1,0 @@
-//set up email content
-const emailRecipients='pujarniharika@gmail.com';
-const emailSubject='timbr: Reminder to water your plant'
-const emailContent='Hello, this is a friendly reminder to water your plant 🌱'
-
-//set up SMS content
-const textRecipients="+13129521148"
-const textBody="timbr: Hello, this is a friendly reminder to water your plant 🌱"
-
-
-
-module.exports={
-  emailRecipients,
-  emailSubject,
-  emailContent,
-  textRecipients,
-  textBody
-}
